refact(sqlmodels): Reorganised according to the new package layout
The code of zms-fastapi will be maintained in separate mobileapp and zmscontent to meet different requirements more flexible - both continue to share the same repo.

zms-addons supersedes zms-datatables and is the repo of the zms.unibe package for ZMS extensions and integrations - instead of the former location at unibe-cms/addons.

Refs:
https://github.com/idasm-unibe-ch/unibe-web-mobile-platform/pull/91#issuecomment-3000027368