refact(agenda): Moved schemas from AddOns to ExternalMethod
This allows to control the mapping of different data sources within one module, (ZMSAgenda.py) which is not only customizable for developers in a DEV environment, but also for ZMSAdministrators in TEST and even PROD/LIVE environments on-the-fly.

This should make it easier to adapt the layout/design template (agenda.zpt) as well as the behaviour on saving the content object (onChangeObjEvt.py).

Please refer also to readme (README.md).