feat(maintenance): Added ZMSSiteOverview
Generates a JSON representation of ZMSSites, which can be cached by a ZMSDataTable and regularly updated by ZMSScheduler. Replaces the overview of ZMSSubDomains for 1st level support, which is enhanced by filtering/searching in zmi.

This JSON representation of ZMSSites may be used as a data source for zms-fastapi without the need of SQL mapping. Furthermore, there are examples to generate and render trees of the ZMSSites.